requirements.txt
python-dotenv
pdf2image
tenacity
langchain
Pillow
ibm-watson-machine-learning